refactor: WeatherViewとWeatherViewEventHandlerを復活
